+++ app/src/main/java/com/xuejian/client/lxp/module/toolbox/im/ContactDetailActivity.java
@@ -183,7 +183,7 @@ public class ContactDetailActivity extends ChatBaseActivity {
         window.setContentView(contentView);
         Display display = windowManager.getDefaultDisplay();
         WindowManager.LayoutParams lp = window.getAttributes();
-        lp.width = (int) (display.getWidth()); // 设置宽度
+        lp.width = display.getWidth(); // 设置宽度
         window.setAttributes(lp);
         window.setGravity(Gravity.BOTTOM); // 此处可以设置dialog显示的位置
         window.setWindowAnimations(R.style.SelectPicDialog); // 添加动画
